fix(systemtags): Make inline tags list fully accessible
1. The label was wrong it mentioned the first and last tag, so the information was duplicated
2. It only mentioned the first and last but not all other
3. All other tags were only added as `title` which is not accessible
