+++ apps/systemtags/src/files_actions/inlineSystemTagsAction.spec.ts
@@ -79,10 +79,8 @@ describe('Inline system tags action render tests', () => {
 
 		const result = await action.renderInline!(file, view)
 		expect(result).toBeInstanceOf(HTMLElement)
-		expect(result!.outerHTML).toBe(
-			'<ul class="files-list__system-tags" aria-label="This file has the tag Confidential">'
-				+ '<li class="files-list__system-tag">Confidential</li>'
-			+ '</ul>',
+		expect(result!.outerHTML).toMatchInlineSnapshot(
+			'"<ul class="files-list__system-tags" aria-label="Assigned collaborative tags"><li class="files-list__system-tag">Confidential</li></ul>"',
 		)
 	})
 
@@ -95,21 +93,15 @@ describe('Inline system tags action render tests', () => {
 			permissions: Permission.ALL,
 			attributes: {
 				'system-tags': {
-					'system-tag': [
-						'Important',
-						'Confidential',
-					],
+					'system-tag': ['Important', 'Confidential'],
 				},
 			},
 		})
 
 		const result = await action.renderInline!(file, view)
 		expect(result).toBeInstanceOf(HTMLElement)
-		expect(result!.outerHTML).toBe(
-			'<ul class="files-list__system-tags" aria-label="This file has the tags Important and Confidential">'
-				+ '<li class="files-list__system-tag">Important</li>'
-				+ '<li class="files-list__system-tag files-list__system-tag--more" title="Confidential">+1</li>'
-			+ '</ul>',
+		expect(result!.outerHTML).toMatchInlineSnapshot(
+			'"<ul class="files-list__system-tags" aria-label="Assigned collaborative tags"><li class="files-list__system-tag">Important</li><li class="files-list__system-tag">Confidential</li></ul>"',
 		)
 	})
 
@@ -134,11 +126,8 @@ describe('Inline system tags action render tests', () => {
 
 		const result = await action.renderInline!(file, view)
 		expect(result).toBeInstanceOf(HTMLElement)
-		expect(result!.outerHTML).toBe(
-			'<ul class="files-list__system-tags" aria-label="This file has the tags Important, Confidential, Secret and Classified">'
-				+ '<li class="files-list__system-tag">Important</li>'
-				+ '<li class="files-list__system-tag files-list__system-tag--more" title="Confidential, Secret, Classified">+3</li>'
-			+ '</ul>',
+		expect(result!.outerHTML).toMatchInlineSnapshot(
+			'"<ul class="files-list__system-tags" aria-label="Assigned collaborative tags"><li class="files-list__system-tag">Important</li><li class="files-list__system-tag files-list__system-tag--more" title="Confidential, Secret, Classified" aria-hidden="true" role="presentation">+3</li><li class="files-list__system-tag hidden-visually">Confidential</li><li class="files-list__system-tag hidden-visually">Secret</li><li class="files-list__system-tag hidden-visually">Classified</li></ul>"',
 		)
 	})
 })
