+++ apps/systemtags/src/files_actions/inlineSystemTagsAction.ts
@@ -2,7 +2,8 @@
  * SPDX-FileCopyrightText: 2023 Nextcloud GmbH and Nextcloud contributors
  * SPDX-License-Identifier: AGPL-3.0-or-later
  */
-import { FileAction, Node, registerDavProperty, registerFileAction } from '@nextcloud/files'
+import type { Node } from '@nextcloud/files'
+import { FileAction } from '@nextcloud/files'
 import { translate as t } from '@nextcloud/l10n'
 
 import '../css/fileEntryInlineSystemTags.scss'
@@ -63,22 +64,30 @@ export const action = new FileAction({
 
 		const systemTagsElement = document.createElement('ul')
 		systemTagsElement.classList.add('files-list__system-tags')
-
-		if (tags.length === 1) {
-			systemTagsElement.setAttribute('aria-label', t('files', 'This file has the tag {tag}', { tag: tags[0] }))
-		} else {
-			const firstTags = tags.slice(0, -1).join(', ')
-			const lastTag = tags[tags.length - 1]
-			systemTagsElement.setAttribute('aria-label', t('files', 'This file has the tags {firstTags} and {lastTag}', { firstTags, lastTag }))
-		}
+		systemTagsElement.setAttribute('aria-label', t('files', 'Assigned collaborative tags'))
 
 		systemTagsElement.append(renderTag(tags[0]))
-
-		// More tags than the one we're showing
-		if (tags.length > 1) {
+		if (tags.length === 2) {
+			// Special case only two tags:
+			// the overflow fake tag would take the same space as this, so render it
+			systemTagsElement.append(renderTag(tags[1]))
+		} else if (tags.length > 1) {
+			// More tags than the one we're showing
+			// So we add a overflow element indicating there are more tags
 			const moreTagElement = renderTag('+' + (tags.length - 1), true)
 			moreTagElement.setAttribute('title', tags.slice(1).join(', '))
+			// because the title is not accessible we hide this element for screen readers (see alternative below)
+			moreTagElement.setAttribute('aria-hidden', 'true')
+			moreTagElement.setAttribute('role', 'presentation')
 			systemTagsElement.append(moreTagElement)
+
+			// For accessibility the tags are listed, as the title is not accessible
+			// but those tags are visually hidden
+			for (const tag of tags.slice(1)) {
+				const tagElement = renderTag(tag)
+				tagElement.classList.add('hidden-visually')
+				systemTagsElement.append(tagElement)
+			}
 		}
 
 		return systemTagsElement
@@ -86,6 +95,3 @@ export const action = new FileAction({
 
 	order: 0,
 })
-
-registerDavProperty('nc:system-tags')
-registerFileAction(action)
